fix(attachment): 예외처리 수정
