+++ src/main/java/com/sparta/springtrello/domain/attachment/exception/GlobalExceptionHandler.java
@@ -1,44 +1,44 @@
-package com.sparta.springtrello.domain.attachment.exception;
-
-
-
-import com.sparta.springtrello.domain.attachment.dto.ErrorResponse;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.multipart.MaxUploadSizeExceededException;
-
-@ControllerAdvice
-public class GlobalExceptionHandler {
-
-    @ExceptionHandler(MaxUploadSizeExceededException.class)
-    public ResponseEntity<ErrorResponse> handleMaxUploadSizeExceededException(MaxUploadSizeExceededException ex) {
-        return ResponseEntity.status(HttpStatus.PAYLOAD_TOO_LARGE)
-                .body(new ErrorResponse(413, "파일 크기가 5MB를 초과합니다.", null));
-    }
-
-    @ExceptionHandler(FileTypeNotSupportedException.class)
-    public ResponseEntity<ErrorResponse> handleFileTypeNotSupportedException(FileTypeNotSupportedException ex) {
-        return ResponseEntity.badRequest()
-                .body(new ErrorResponse(400, ex.getMessage(), null));
-    }
-
-    @ExceptionHandler(InvalidRequestException.class)
-    public ResponseEntity<ErrorResponse> handleInvalidRequestException(InvalidRequestException ex) {
-        return ResponseEntity.status(HttpStatus.NOT_FOUND)
-                .body(new ErrorResponse(404, ex.getMessage(), null));
-    }
-
-    @ExceptionHandler(AttachmentNotFoundException.class)
-    public ResponseEntity<ErrorResponse> handleAttachmentNotFoundException(AttachmentNotFoundException ex) {
-        return ResponseEntity.status(HttpStatus.NOT_FOUND)
-                .body(new ErrorResponse(404, "해당 카드 ID에 대한 첨부파일이 없습니다.", null));
-    }
-
-    @ExceptionHandler(Exception.class)
-    public ResponseEntity<ErrorResponse> handleGeneralException(Exception ex) {
-        return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
-                .body(new ErrorResponse(500, "서버 오류 발생", null));
-    }
-}
+//package com.sparta.springtrello.domain.attachment.exception;
+//
+//
+//
+//import com.sparta.springtrello.domain.attachment.dto.ErrorResponse;
+//import org.springframework.http.HttpStatus;
+//import org.springframework.http.ResponseEntity;
+//import org.springframework.web.bind.annotation.ControllerAdvice;
+//import org.springframework.web.bind.annotation.ExceptionHandler;
+//import org.springframework.web.multipart.MaxUploadSizeExceededException;
+//
+//@ControllerAdvice
+//public class GlobalExceptionHandler {
+//
+//    @ExceptionHandler(MaxUploadSizeExceededException.class)
+//    public ResponseEntity<ErrorResponse> handleMaxUploadSizeExceededException(MaxUploadSizeExceededException ex) {
+//        return ResponseEntity.status(HttpStatus.PAYLOAD_TOO_LARGE)
+//                .body(new ErrorResponse(413, "파일 크기가 5MB를 초과합니다.", null));
+//    }
+//
+//    @ExceptionHandler(FileTypeNotSupportedException.class)
+//    public ResponseEntity<ErrorResponse> handleFileTypeNotSupportedException(FileTypeNotSupportedException ex) {
+//        return ResponseEntity.badRequest()
+//                .body(new ErrorResponse(400, ex.getMessage(), null));
+//    }
+//
+//    @ExceptionHandler(InvalidRequestException.class)
+//    public ResponseEntity<ErrorResponse> handleInvalidRequestException(InvalidRequestException ex) {
+//        return ResponseEntity.status(HttpStatus.NOT_FOUND)
+//                .body(new ErrorResponse(404, ex.getMessage(), null));
+//    }
+//
+//    @ExceptionHandler(AttachmentNotFoundException.class)
+//    public ResponseEntity<ErrorResponse> handleAttachmentNotFoundException(AttachmentNotFoundException ex) {
+//        return ResponseEntity.status(HttpStatus.NOT_FOUND)
+//                .body(new ErrorResponse(404, "해당 카드 ID에 대한 첨부파일이 없습니다.", null));
+//    }
+//
+//    @ExceptionHandler(Exception.class)
+//    public ResponseEntity<ErrorResponse> handleGeneralException(Exception ex) {
+//        return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
+//                .body(new ErrorResponse(500, "서버 오류 발생", null));
+//    }
+//}
